@@ -1,11 +1,12 @@
 source 'https://rubygems.org'
 
-ruby '>= 2.2.2'
+ruby '>= 2.4.1'
 
 # ------------------------------------------------
 #    RAILS
-gem 'rails', '4.2.7'
-gem 'railties', '~> 4.2'
+gem 'rails', '~> 4.2.10'
+gem 'railties'
+
 
 #    GEMS ADDED TO HELP HANDLE RAILS MIGRATION FROM 3.x to 4.2
 #    THESE GEMS HELP SUPPORT DEPRACATED FUNCTIONALITY AND WILL LOSE SUPPORT IN FUTURE VERSIONS
