@@ -13,31 +13,31 @@ module Api
 
         @org_templates = {}
 
-        published_templates = Template.includes(:org).valid.where(customization_of: nil, published: true).order(:org_id, :version)
-        customized_templates = Template.includes(:org).valid.where(org_id: @user.org_id, published: true).where.not(customization_of: nil)
+        published_templates = Template.includes(:org).unarchived.where(customization_of: nil, published: true).order(:org_id, :version)
+        customized_templates = Template.includes(:org).unarchived.where(org_id: @user.org_id, published: true).where.not(customization_of: nil)
 
-        published_templates.each do |temp|
+        Template.published.order(:org_id, :version).each do |temp|
           if @org_templates[temp.org].present?
-            if @org_templates[temp.org][:own][temp.dmptemplate_id].nil?
-              @org_templates[temp.org][:own][temp.dmptemplate_id] = temp
+            if @org_templates[temp.org][:own][temp.family_id].nil?
+              @org_templates[temp.org][:own][temp.family_id] = temp
             end
           else
             @org_templates[temp.org] = {}
             @org_templates[temp.org][:own] = {}
             @org_templates[temp.org][:cust] = {}
-            @org_templates[temp.org][:own][temp.dmptemplate_id] = temp
+            @org_templates[temp.org][:own][temp.family_id] = temp
           end
         end
-        customized_templates.each do |temp|
+        Template.published_customization(@user.org_id).each do |temp|
           if @org_templates[temp.org].present?
-            if @org_templates[temp.org][:cust][temp.dmptemplate_id].nil?
-              @org_templates[temp.org][:cust][temp.dmptemplate_id] = temp
+            if @org_templates[temp.org][:cust][temp.family_id].nil?
+              @org_templates[temp.org][:cust][temp.family_id] = temp
             end
           else
             @org_templates[temp.org] = {}
             @org_templates[temp.org][:own] = {}
             @org_templates[temp.org][:cust] = {}
-            @org_templates[temp.org][:cust][temp.dmptemplate_id] = temp
+            @org_templates[temp.org][:cust][temp.family_id] = temp
           end
         end
         respond_with @org_templates
