@@ -1,79 +1,65 @@
 class Paginable::TemplatesController < ApplicationController
   include Paginable
-  include TemplateFilter
       
-  # GET /paginable/templates/all/:page  (AJAX)
+  # GET /paginable/templates/:page  (AJAX)
   # -----------------------------------------------------
-  def all
-    raise Pundit::NotAuthorizedError unless Paginable::TemplatePolicy.new(current_user).all?
-    # Apply scoping
-    hash = apply_scoping(params[:scope] || 'all', false, true)
-
-    # Apply pagination
-    hash[:templates] = hash[:templates].page(params[:page]) if params[:page] != 'ALL'
-    
-    # Gather up all of the publication dates for the live versions of each template.
-    published = get_publication_dates(hash[:scopes][:dmptemplate_ids])
-    
-    paginable_renderise partial: 'all',
-                        scope: hash[:templates],
-                        locals: { current_org: current_user.org.id,
-                                  customizations: hash[:customizations],
-                                  published: published,
-                                  scopes: hash[:scopes]}
+  def index
+    authorize Template
+    templates = Template.latest_version.where(customization_of: nil)
+    case params[:f]
+    when 'published'
+      template_ids = templates.select{|t| t.published? || t.draft? }.collect(&:family_id)
+      templates = Template.latest_version(template_ids).where(customization_of: nil)
+    when 'unpublished'
+      template_ids = templates.select{|t| !t.published? && !t.draft? }.collect(&:family_id)
+      templates = Template.latest_version(template_ids).where(customization_of: nil)
+    end
+    paginable_renderise partial: 'index', scope: templates.includes(:org), locals: { action: 'index' }
   end
   
-  # GET /paginable/templates/funders/:page  (AJAX)
+  # GET /paginable/templates/organisational/:page  (AJAX)
   # -----------------------------------------------------
-  def funders
-    raise Pundit::NotAuthorizedError unless Paginable::TemplatePolicy.new(current_user).funders?
-    # Apply scoping
-    hash = apply_scoping(params[:scope] || 'all', true, false)
-
-    # Apply pagination
-    hash[:templates] = hash[:templates].page(params[:page]) if params[:page] != 'ALL'
-    
-    # Gather up all of the publication dates for the live versions of each template.
-    published = get_publication_dates(hash[:scopes][:dmptemplate_ids])
-    
-    paginable_renderise partial: 'funders',
-                        scope: hash[:templates],
-                        locals: { current_org: current_user.org.id,
-                                  customizations: hash[:customizations],
-                                  published: published,
-                                  scopes: hash[:scopes] }
+  def organisational
+    authorize Template
+    templates = Template.latest_version_per_org(current_user.org.id).where(customization_of: nil, org_id: current_user.org.id)
+    case params[:f]
+    when 'published'
+      template_ids = templates.select{|t| t.published? || t.draft? }.collect(&:family_id)
+      templates = Template.latest_version(template_ids)
+    when 'unpublished'
+      template_ids = templates.select{|t| !t.published? && !t.draft? }.collect(&:family_id)
+      templates = Template.latest_version(template_ids)
+    end
+    paginable_renderise partial: 'organisational', scope: templates, locals: { action: 'organisational' }
   end
   
-  # GET /paginable/templates/orgs/:page  (AJAX)
+  # GET /paginable/templates/customisable/:page  (AJAX)
   # -----------------------------------------------------
-  def orgs
-    raise Pundit::NotAuthorizedError unless Paginable::TemplatePolicy.new(current_user).orgs?
-    # Apply scoping
-    hash = apply_scoping(params[:scope] || 'all', false, false)
-    
-    # Apply pagination
-    hash[:templates] = hash[:templates].page(params[:page]) if params[:page] != 'ALL'
-    
-    # Gather up all of the publication dates for the live versions of each template.
-    published = get_publication_dates(hash[:scopes][:dmptemplate_ids])
-    
-    paginable_renderise partial: 'orgs',
-                        scope: hash[:templates],
-                        locals: { current_org: current_user.org.id,
-                                  customizations: hash[:customizations],
-                                  published: published,
-                                  scopes: hash[:scopes]}
+  def customisable
+    authorize Template
+    customizations = Template.latest_customized_version_per_org(current_user.org.id)
+    templates = Template.latest_customizable
+    case params[:f]
+      when 'published'
+        customization_ids = customizations.select{|t| t.published? || t.draft? }.collect(&:customization_of)
+        templates = Template.latest_customizable.where(family_id: customization_ids)
+      when 'unpublished'
+        customization_ids = customizations.select{|t| !t.published? && !t.draft? }.collect(&:customization_of)
+        templates = Template.latest_customizable.where(family_id: customization_ids)
+      when 'not-customised'
+        templates = Template.latest_customizable.where.not(family_id: customizations.collect(&:customization_of))
+    end
+    paginable_renderise partial: 'customisable', scope: templates.joins(:org), locals: { action: 'customisable', customizations: customizations }
   end
 
   # GET /paginable/templates/publicly_visible/:page  (AJAX)
   # -----------------------------------------------------
   def publicly_visible
-    templates = Template.live(Template.families(Org.funder.pluck(:id)).pluck(:dmptemplate_id)).publicly_visible.pluck(:id) <<
-    Template.where(is_default: true).valid.published.pluck(:id)
-    
+    templates = Template.live(Template.families(Org.funder.pluck(:id)).pluck(:family_id)).publicly_visible.pluck(:id) <<
+    Template.where(is_default: true).unarchived.published.pluck(:id)
     paginable_renderise(
       partial: 'publicly_visible',
-      scope: Template.includes(:org).where(id: templates.uniq.flatten).valid.published)
+      scope: Template.includes(:org).where(id: templates.uniq.flatten).published)
   end
 
   # GET /paginable/templates/:id/history/:page  (AJAX)
@@ -81,12 +67,11 @@ class Paginable::TemplatesController < ApplicationController
   def history
     @template = Template.find(params[:id])
     authorize @template
-    @templates = Template.where(dmptemplate_id: @template.dmptemplate_id)
-    @current = Template.current(@template.dmptemplate_id)
+    @templates = Template.where(family_id: @template.family_id)
+    @current = Template.current(@template.family_id)
     paginable_renderise(
       partial: 'history',
       scope: @templates,
-      query_params: { sort_field: :version, sort_direction: :desc },
-      locals: { current: @current })
+      locals: { current: templates.maximum(:version) })
   end
 end
