@@ -147,11 +147,35 @@ class PlansController < ApplicationController
     @all_ggs_grouped_by_org = @all_ggs_grouped_by_org.sort_by {|org,gg| (org.nil? ? '' : org.name)}
     @selected_guidance_groups = @selected_guidance_groups.collect{|gg| gg.id}
 
-    @based_on = (@plan.template.customization_of.nil? ? @plan.template : Template.where(dmptemplate: @plan.template.customization_of).first)
+    @based_on = (@plan.template.customization_of.nil? ? @plan.template : Template.where(family_id: @plan.template.customization_of).first)
 
     respond_to :html
   end
 
+  # GET /plans/:plan_id/phases/:id/edit
+  def edit
+    plan = Plan.find(params[:id])
+    authorize plan
+    
+    plan, phase = Plan.load_for_phase(params[:id], params[:phase_id])
+    
+    readonly = !plan.editable_by?(current_user.id)
+    
+    guidance_groups_ids = plan.guidance_groups.collect(&:id)
+    
+    guidance_groups =  GuidanceGroup.where(published: true, id: guidance_groups_ids)
+    # Since the answers have been pre-fetched through plan (see Plan.load_for_phase)
+    # we create a hash whose keys are question id and value is the answer associated
+    answers = plan.answers.reduce({}){ |m, a| m[a.question_id] = a; m }
+    
+    render('/phases/edit', locals: {
+      base_template_org: phase.template.base_org,
+      plan: plan, phase: phase, readonly: readonly,
+      question_guidance: plan.guidance_by_question_as_hash,
+      guidance_groups: guidance_groups,
+      answers: answers })
+  end
+  
   # PUT /plans/1
   # PUT /plans/1.json
   def update
@@ -364,13 +388,13 @@ class PlansController < ApplicationController
   end
 
 
-  # different versions of the same template have the same dmptemplate_id
+  # different versions of the same template have the same family_id
   # but different version numbers so for each set of templates with the
-  # same dmptemplate_id choose the highest version number.
+  # same family_id choose the highest version number.
   def get_most_recent( templates )
     groups = Hash.new
     templates.each do |t|
-      k = t.dmptemplate_id
+      k = t.family_id
       if !groups.has_key?(k)
         groups[k] = t
       else
