@@ -4,16 +4,18 @@ class Section < ActiveRecord::Base
   belongs_to :phase
   belongs_to :organisation
   has_many :questions, :dependent => :destroy
+  has_one :template, through: :phase
 
   #Link the data
   accepts_nested_attributes_for :questions, :reject_if => lambda {|a| a[:text].blank? },  :allow_destroy => true
-#  accepts_nested_attributes_for :version
 
   attr_accessible :phase_id, :description, :number, :title, :published,
                   :questions_attributes, :organisation, :phase, :modifiable,
                   :as => [:default, :admin]
 
   validates :phase, :title, :number, presence: {message: _("can't be blank")}
+
+  before_validation :set_defaults
 
   ##
   # return the title of the section
@@ -35,19 +37,18 @@ class Section < ActiveRecord::Base
     end
   end
 
-  ##
-  # deep copy of the given section and all it's associations
-  #
-  # @params [Section] section to be deep copied
-  # @return [Section] the saved, copied section
-  def self.deep_copy(section)
-    section_copy = section.dup
-    section_copy.save!
-    section.questions.each do |question|
-      question_copy = Question.deep_copy(question)
-      question_copy.section_id = section_copy.id
-      question_copy.save!
-    end
-    return section_copy
+  def deep_copy(**options)
+    copy = self.dup
+    copy.modifiable = options.fetch(:modifiable, self.modifiable)
+    copy.phase_id = options.fetch(:phase_id, nil)
+    copy.save!(validate: false)  if options.fetch(:save, false)
+    options[:section_id] = id
+    self.questions.map{ |question| copy.questions << question.deep_copy(options) }
+    return copy
   end
+
+  private
+    def set_defaults
+      self.modifiable = true if modifiable.nil?
+    end
 end
