@@ -1,9 +1,11 @@
 class Template < ActiveRecord::Base
   include GlobalHelpers
   include ActiveModel::Validations
+  include TemplateScope
   validates_with TemplateLinksValidator
 
-  before_validation :set_creation_defaults
+  before_validation :set_defaults 
+  after_update :reconcile_published, if: Proc.new { |template| template.published? }
 
   # Stores links as an JSON object: { funder: [{"link":"www.example.com","text":"foo"}, ...], sample_plan: [{"link":"www.example.com","text":"foo"}, ...]}
   # The links is validated against custom validator allocated at validators/template_links_validator.rb
@@ -17,15 +19,12 @@ class Template < ActiveRecord::Base
   has_many :sections, through: :phases
   has_many :questions, through: :sections
 
-  has_many :customizations, class_name: 'Template', foreign_key: 'dmptemplate_id'
-  belongs_to :dmptemplate, class_name: 'Template'
-
   ##
   # Possibly needed for active_admin
   #   -relies on protected_attributes gem as syntax depricated in rails 4.2
   attr_accessible :id, :org_id, :description, :published, :title, :locale, :customization_of,
-                  :is_default, :guidance_group_ids, :org, :plans, :phases, :dmptemplate_id,
-                  :migrated, :version, :visibility, :published, :links, :as => [:default, :admin]
+                  :is_default, :guidance_group_ids, :org, :plans, :phases, :family_id,
+                  :archived, :version, :visibility, :published, :links, :as => [:default, :admin]
 
   # A standard template should be organisationally visible. Funder templates that are 
   # meant for external use will be publicly visible. This allows a funder to create 'funder' as
@@ -37,262 +36,217 @@ class Template < ActiveRecord::Base
     s.key :export, defaults: Settings::Template::DEFAULT_SETTINGS
   end
 
-  validates :org, :title, :version, presence: {message: _("can't be blank")}
+  validates :org, :title, presence: {message: _("can't be blank")}
 
-  scope :valid,  -> { where(migrated: false) }
-  scope :published, -> { where(published: true) }
-
-  # Retrieves all valid and published templates
-  scope :valid_published, -> (is_default: false)  {
-    Template.where(templates: { is_default: is_default }).valid().published()
-  }
-
-  scope :publicly_visible, -> { where(:visibility => Template.visibilities[:publicly_visible]) }
-  scope :organisationally_visible, -> { where(:visibility => Template.visibilities[:organisationally_visible]) }
-  
-  # Retrieves template with distinct dmptemplate_id that are valid (e.g. migrated false) and customization_of is nil. Note,
-  # if organisation ids are passed, the query will filter only those distinct dmptemplate_ids for those organisations
-  scope :families, -> (org_ids=nil) {
-    if org_ids.is_a?(Array) 
-      valid.where(org_id: org_ids, customization_of: nil).distinct
-    else
-      valid.where(customization_of: nil).distinct
-    end 
-  }
-  # Retrieves the maximum version for the array of dmptemplate_ids passed. If dmptemplate_ids is missing, every maximum
-  # version for each different dmptemplate_id will be retrieved
-  scope :dmptemplate_ids_with_max_version, -> (dmptemplate_ids=nil) {
-    if dmptemplate_ids.is_a?(Array)
-      select("MAX(version) AS version", :dmptemplate_id).where(dmptemplate_id: dmptemplate_ids).group(:dmptemplate_id)
-    else
-      select("MAX(version) AS version", :dmptemplate_id).group(:dmptemplate_id)
+  # Class methods gets defined within this 
+  class << self
+    def current(family_id)
+      unarchived.where(family_id: family_id).order(version: :desc).first
     end
-  }
-  # Retrieves the maximum version for the array of customization_ofs passed. If customization_ofs is missing, every maximum
-  # version for each different customization_of will be retrieved
-  scope :customization_ofs_with_max_version, -> (customization_ofs=nil, org_id=nil) {
-    chained_scope = select("MAX(version) AS version", :customization_of)
-    if customization_ofs.respond_to?(:each)
-      chained_scope = chained_scope.where(customization_of: customization_ofs)
-    end
-    if org_id.present?
-      chained_scope = chained_scope.where(org_id: org_id)
-    end
-    chained_scope.group(:customization_of)
-  }
-  # Retrieves the latest template version, i.e. the one with maximum version for each dmptemplate_id
-  scope :latest_version, -> (dmptemplate_ids=nil) {
-    from(dmptemplate_ids_with_max_version(dmptemplate_ids), :current)
-    .joins("INNER JOIN templates ON current.version = templates.version"\
-      " AND current.dmptemplate_id = templates.dmptemplate_id")
-  }
-  # Retrieves the latest customized version, i.e. the one with maximum version for each customization_of=dmptemplate_id
-  scope :latest_customization, -> (org_id, dmptemplate_ids=nil) {
-    from(customization_ofs_with_max_version(dmptemplate_ids, org_id), :current)
-    .joins("INNER JOIN templates ON current.version = templates.version"\
-      " AND current.customization_of = templates.customization_of")
-    .where('templates.org_id = ?', org_id)
-  }
-  
-  scope :search, -> (term) {
-    search_pattern = "%#{term}%"
-    joins(:org).where("templates.title LIKE ? OR orgs.name LIKE ?", search_pattern, search_pattern)
-  }
-  
-  # Retrieves the list of all dmptemplate_ids (template versioning families) for the specified Org
-  def self.dmptemplate_ids
-    Template.all.valid.distinct.pluck(:dmptemplate_id)
-  end
-
-  # Retrieves the most recent version of the template for the specified dmptemplate_id
-  def self.current(dmptemplate_id)
-    Template.where(dmptemplate_id: dmptemplate_id).order(version: :desc).valid.first
-  end
-
-  # Retrieves the current published version of the template for the specified Org and dmptemplate_id
-  def self.live(dmptemplate_id)
-    if dmptemplate_id.respond_to?(:each)
-      Template.where(dmptemplate_id: dmptemplate_id, published: true).valid
-    else
-      Template.where(dmptemplate_id: dmptemplate_id, published: true).valid.first
-    end
-  end
-
-  def self.default
-    Template.valid.where(is_default: true, published: true).order(:version).last
-  end
-
-  ##
-  # Retrieves the most current customization of the template for the
-  # specified org and dmptemplate_id
-  # returns nil if no customizations found
-  #
-  # @params  dmptemplate_ids of the original template
-  # @params [integer] org_id for the customizing organisation
-  # @return [nil, Template] the customized template or nil
-  def self.org_customizations(dmptemplate_ids, org_id)
-    template_ids = latest_customization(org_id, dmptemplate_ids).pluck(:id)
-    includes(:org).where(id: template_ids)
-  end
-  
-  # Retrieves current templates with their org associated for a set of valid orgs
-  # TODO pass an array of org ids instead of Org instances
-  def self.get_latest_template_versions(orgs)
-    if orgs.respond_to?(:each)
-      families_ids = families(orgs.map(&:id)).pluck(:dmptemplate_id)
-    elsif orgs.is_a?(Org)
-      families_ids = families([orgs.id]).pluck(:dmptemplate_id)
-    else
-      families_ids = []
-    end
-    template_ids = latest_version(families_ids).pluck(:id)
-    includes(:org).where(id: template_ids)
-  end
-  
-  # Retrieves current templates with their org associated for a set of valid orgs
-  # TODO pass an array of org ids instead of Org instances
-  def self.get_public_published_template_versions(orgs)
-    if orgs.respond_to?(:each)
-      families_ids = families(orgs.map(&:id)).pluck(:dmptemplate_id)
-    elsif orgs.is_a?(Org)
-      families_ids = families([orgs.id]).pluck(:dmptemplate_id)
-    else
-      families_ids = []
-    end
-    includes(:org).where(dmptemplate_id: families_ids, published: true, visibility: Template.visibilities[:publicly_visible])
-  end
-  
-  ##
-  # create a new version of the most current copy of the template
-  #
-  # @return [Template] new version
-  def get_new_version
-    if self.id.present?
-      new_version = Template.deep_copy(self)
-      new_version.version = (self.version + 1)
-      new_version.published = false 
-      new_version.visibility = self.visibility # do not change the visibility 
-      new_version.is_default = self.is_default # retain the default template flag
-      new_version.save!
-      new_version
-    else
-      nil
-    end
-  end
-  
-  ##
-  # deep copy the given template and all of it's associations
-  #
-  # @params [Template] template to be deep copied
-  # @return [Template] saved copied template
-  def self.deep_copy(template)
-    template_copy = template.dup
-    template_copy.save!
-    template.phases.each do |phase|
-      phase_copy = Phase.deep_copy(phase)
-      phase_copy.template_id = template_copy.id
-      phase_copy.save!
-    end
-    return template_copy
-  end
-
-
-  # EVALUATE CLASS AND INSTANCE METHODS BELOW
-  #
-  # What do they do? do they do it efficiently, and do we need them?
-
-
-  ##
-  # convert the given template to a hash and return with all it's associations
-  # to use, please pre-fetch org, phases, section, questions, annotations,
-  #   question_options, question_formats,
-  # TODO: Themes & guidance?
-  #
-  # @return [hash] hash of template, phases, sections, questions, question_options, annotations
-  def to_hash
-    hash = {}
-    hash[:template] = {}
-    hash[:template][:data] = self
-    hash[:template][:org] = self.org
-    phases = {}
-    hash[:template][:phases] = phases
-    self.phases.each do |phase|
-      phases[phase.number] = {}
-      phases[phase.number][:data] = phase
-      phases[phase.number][:sections] = {}
-      phase.sections.each do |section|
-        phases[phase.number][:sections][section.number] = {}
-        phases[phase.number][:sections][section.number][:data] = section
-        phases[phase.number][:sections][section.number][:questions] = {}
-        section.questions.each do |question|
-          phases[phase.number][:sections][section.number][:questions][question.number] = {}
-          phases[phase.number][:sections][section.number][:questions][question.number][:data] = question
-          phases[phase.number][:sections][section.number][:questions][question.number][:annotations] = {}
-          question.annotations.each do |annotation|
-            phases[phase.number][:sections][section.number][:questions][question.number][:annotations][annotation.id] = {}
-            phases[phase.number][:sections][section.number][:questions][question.number][:annotations][annotation.id][:data] = annotation
-          end
-          phases[phase.number][:sections][section.number][:questions][question.number][:question_options] = {}
-          question.question_options.each do |question_option|
-            phases[phase.number][:sections][section.number][:questions][question.number][:question_options][:data] = question_option
-            phases[phase.number][:sections][section.number][:questions][question.number][:question_format] = question.question_format
-          end
-        end
+    def live(family_id)
+      if family_id.respond_to?(:each)
+        unarchived.where(family_id: family_id, published: true)
+      else
+        unarchived.where(family_id: family_id, published: true).first
       end
     end
-    return hash
+    def find_or_generate_version!(template)
+      if template.latest?
+        if template.generate_version?
+          return template.generate_version!
+        end
+        return template
+      end
+      raise _('A historical template cannot be retrieved for being modified')
+    end
   end
 
-# TODO: Why are we passing in an org and template here?
-  ##
-  # Verify if a template has customisation by given organisation
-  #
-  # @param org_id [integer] the integer id for an organisation
-  # @param temp [dmptemplate] a template object
-  # @return [Boolean] true if temp has customisation by the given organisation
-  def has_customisations?(org_id, temp)
-    modifiable = true
-    phases.each do |phase|
-      modifiable = modifiable && phase.modifiable
+  # Creates a copy of the current template
+  # raises ActiveRecord::RecordInvalid when save option is true and validations fails
+  def deep_copy(attributes: {}, **options)
+    copy = self.dup
+    if attributes.respond_to?(:each_pair)
+      attributes.each_pair{ |attribute, value| copy.send("#{attribute}=".to_sym, value) if copy.respond_to?("#{attribute}=".to_sym) }
     end
-    return !modifiable
-  end
-  
-  def template_type
-    self.customization_of.present? ? _('customisation') : _('template')
+    copy.save! if options.fetch(:save, false)
+    options[:template_id] = copy.id
+    self.phases.each{ |phase| copy.phases << phase.deep_copy(options) }
+    return copy
   end
 
   # Retrieves the template's org or the org of the template this one is derived
   # from of it is a customization
   def base_org
     if self.customization_of.present?
-      base_template_org = Template.where(dmptemplate_id: self.customization_of).first.org
+      return Template.where(family_id: self.customization_of).first.org
     else
-      base_template_org = self.org
+      return self.org
     end
   end
 
-  # --------------------------------------------------------
-  private
-  # Initialize the published and dirty flags for new templates
-  def set_creation_defaults
-    # Only run this before_validation because rails fires this before save/create
-    if self.id.nil?
-      self.published = false
-      self.migrated = false
-      self.dirty = false
-      self.is_default = false if self.is_default.nil?
-      self.version = 0 if self.version.nil?
-      # Organisationally visible by default unless Org is only a funder
-      self.visibility = (self.org.present? && self.org.funder_only?) ? Template.visibilities[:publicly_visible] : Template.visibilities[:organisationally_visible] 
-      
-      # Generate a unique identifier for the dmptemplate_id if necessary
-      if self.dmptemplate_id.nil?
-        self.dmptemplate_id = loop do
-          random = rand 2147483647
-          break random unless Template.exists?(dmptemplate_id: random)
+  # Returns whether or not this is the latest version of the current template's family
+  def latest?
+    return (self.id == Template.latest_version(self.family_id).pluck(:id).first)
+  end
+  # Determines whether or not a new version should be generated
+  def generate_version?
+    return self.published
+  end
+  # Determines whether or not a customization for the customizing_org passed should be generated
+  def customize?(customizing_org)
+    if customizing_org.is_a?(Org) && (self.org.funder_only? || self.is_default)
+      return !Template.unarchived.where(customization_of: self.family_id, org: customizing_org).exists?
+    end
+    return false
+  end
+  # Determines whether or not a customized template should be upgraded
+  def upgrade_customization?
+    if self.customization_of.present?
+      funder_template = Template.published(self.customization_of).select(:created_at).first
+      if funder_template.present?
+        return funder_template.created_at > self.created_at
+      end
+    end
+    return false
+  end
+
+  # Checks to see if the template family has a published version and if its not the current template
+  def draft?
+    return !self.published && Template.published(self.family_id).length > 0
+  end
+
+  def removable?
+    versions = Template.includes(:plans).where(family_id: self.family_id)
+    return versions.select{|version| version.plans.length > 0 }.empty?
+  end
+
+  # Returns a new unpublished copy of self with a new family_id, version = zero for the specified org
+  def generate_copy!(org)
+    raise _('generate_copy! requires an organisation target') unless org.is_a?(Org) # Assume customizing_org is persisted
+    template = deep_copy(
+      attributes: {
+        version: 0,
+        published: false,
+        family_id: new_family_id,
+        org: org,
+        is_default: false,
+        title: _('Copy of %{template}') % { template: self.title }
+      }, modifiable: true, save: true)
+    return template
+  end
+
+  # Generates a new copy of self with an incremented version number
+  def generate_version!
+    raise _('generate_version! requires a published template') unless published
+    template = deep_copy(
+      attributes: {
+        version: self.version+1,
+        published: false,
+        org: self.org
+      }, save: true)
+    return template
+  end
+
+  # Generates a new copy of self for the specified customizing_org
+  def customize!(customizing_org)
+    raise _('customize! requires an organisation target') unless customizing_org.is_a?(Org) # Assume customizing_org is persisted
+    raise _('customize! requires a template from a funder') if !self.org.funder_only? && !self.is_default # Assume self has org associated
+    customization = deep_copy(
+      attributes: {
+        version: 0,
+        published: false,
+        family_id: new_family_id,
+        customization_of: self.family_id,
+        org: customizing_org,
+        visibility: Template.visibilities[:organisationally_visible],
+        is_default: false
+      }, modifiable: false, save: true)
+    return customization
+  end
+  
+  # Generates a new copy of self including latest changes from the funder this template is customized_of
+  def upgrade_customization!
+    raise _('upgrade_customization! requires a customised template') unless customization_of.present?
+    funder_template = Template.published(self.customization_of).first
+    raise _('upgrade_customization! cannot be carried out since there is no published template of its current funder') unless funder_template.present?
+    source = deep_copy(attributes: { version: self.version+1, published: false }) # preserves modifiable flags from the self template copied
+    # Creates a new customisation for the published template whose family_id is self.customization_of
+    customization = funder_template.deep_copy(
+      attributes: {
+        version: source.version,
+        published: source.published,
+        family_id: source.family_id,
+        customization_of: source.customization_of,
+        org: source.org,
+        visibility: Template.visibilities[:organisationally_visible],
+        is_default: false
+      }, modifiable: false, save: true)
+    # Sorts the phases from the source template, i.e. self
+    sorted_phases = source.phases.sort{ |phase1,phase2| phase1.number <=> phase2.number }
+    # Merges modifiable sections or questions from source into customization template object
+    customization.phases.each do |customization_phase|
+      # Search for the phase in the source template whose number matches the customization_phase
+      candidate_phase = sorted_phases.bsearch{ |phase| customization_phase.number <=> phase.number }
+      if candidate_phase.present? # The funder could have added this new phase after the customisation took place
+        # Selects modifiable sections from the candidate_phase
+        modifiable_sections = candidate_phase.sections.select{ |section| section.modifiable }
+        # Attaches modifiable sections into the customization_phase
+        modifiable_sections.each{ |modifiable_section| customization_phase.sections << modifiable_section }
+        # Sorts the sections for the customization_phase
+        sorted_sections = customization_phase.sections.sort{ |section1, section2| section1.number <=> section2.number }
+        # Selects unmodifiable sections from the candidate_phase
+        unmodifiable_sections = candidate_phase.sections.select{ |section| !section.modifiable }
+        unmodifiable_sections.each do |unmodifiable_section|
+          # Search for modifiable questions within the unmodifiable_section from candidate_phase
+          modifiable_questions = unmodifiable_section.questions.select{ |question| question.modifiable }
+          customization_section = sorted_sections.bsearch{ |section| unmodifiable_section.number <=> section.number }
+          if customization_section.present? # The funder could have deleted the section
+            modifiable_questions.each{ |modifiable_question| customization_section.questions << modifiable_question; }
+          end
+          # Search for unmodifiable questions within the unmodifiable_section in case source template added annotations
+          unmodifiable_questions = unmodifiable_section.questions.select{ |question| !question.modifiable }
+          sorted_questions = customization_section.questions.sort{ |question1, question2| question1.number <=> question2.number }
+          unmodifiable_questions.each do |unmodifiable_question|
+            customization_question = sorted_questions.bsearch{ |question| unmodifiable_question.number <=> question.number }
+            if customization_question.present?  # The funder could have deleted the question
+              annotations_added_by_customiser = unmodifiable_question.annotations.select{ |annotation| annotation.org_id == source.org_id }
+              annotations_added_by_customiser.each{ |annotation| customization_question.annotations << annotation }
+            end
+          end
         end
       end
     end
+    # Appends the modifiable phases from source
+    source.phases.select{ |phase| phase.modifiable }.each{ |modifiable_phase| customization.phases << modifiable_phase }
+    return customization
   end
+
+  private
+    # Generate a new random family identifier
+    def new_family_id
+      family_id = loop do
+        random = rand 2147483647
+        break random unless Template.exists?(family_id: random)
+      end
+      family_id
+    end
+    
+    # Default values to set before running any validation
+    def set_defaults
+      self.published ||= false
+      self.archived ||= false
+      self.is_default ||= false
+      self.version ||= 0
+      self.visibility = ((self.org.present? && self.org.funder_only?) || self.is_default?) ? Template.visibilities[:publicly_visible] : Template.visibilities[:organisationally_visible] unless self.id.present?
+      self.customization_of ||= nil
+      self.family_id ||= new_family_id
+      self.archived ||= false
+      self.links ||= { funder: [], sample_plan: [] }
+    end
+    
+    # Only one version of a template should be published at a time, so if this one was published make sure other versions are not
+    def reconcile_published
+      # Unpublish all other versions of this template family
+      Template.where('family_id = ? AND published = ? AND id != ?', self.family_id, true, self.id).update_all(published: false)
+    end
 end
