@@ -13,15 +13,31 @@ class SectionPolicy < ApplicationPolicy
   #  - The template which they are modifying belongs to their org
   ##
 
-  def admin_create?
+  def index?
+    user.present?
+  end
+  
+  def show?
+    user.present?
+  end
+
+  def edit?
+    user.can_modify_templates?  &&  (section.phase.template.org_id == user.org_id)
+  end
+    
+  def new?
     user.can_modify_templates?  &&  (section.phase.template.org_id == user.org_id)
   end
 
-  def admin_update?
+  def create?
     user.can_modify_templates?  &&  (section.phase.template.org_id == user.org_id)
   end
 
-  def admin_destroy?
+  def update?
+    user.can_modify_templates?  &&  (section.phase.template.org_id == user.org_id)
+  end
+
+  def destroy?
     user.can_modify_templates?  &&  (section.phase.template.org_id == user.org_id)
   end
 
