@@ -1,9 +1,8 @@
 class ChangeLongStringsToText < ActiveRecord::Migration
   def change
-    change_column :orgs, :links, :text
-    change_column :templates, :links, :text
-    change_column :identifier_schemes, :logo_url, :text
-    change_column :identifier_schemes, :user_landing_url, :text
+    # change_column :orgs, :links, :text
+    # change_column :templates, :links, :text
+    # change_column :identifier_schemes, :logo_url, :text
+    # change_column :identifier_schemes, :user_landing_url, :text
   end
 end
-
