@@ -45,7 +45,13 @@ export default () => {
         const panelCollapse = $(el);
         // Expands or collapses the panel according to the
         // direction passed (e.g. show --> expands, hide --> collapses)
-        $(el).collapse(direction);
+        if (direction === 'show') {
+          if (!panelCollapse.hasClass('in')) {
+            panelCollapse.prev().trigger('click');
+          }
+        } else {
+          panelCollapse.collapse(direction);
+        }
         // Sets icon at panel-title accordingly
         panelCollapse.prev().find('i.fa')
           .removeClass('fa-plus fa-minus').addClass(direction === 'show' ? 'fa-minus' : 'fa-plus');
